django_backend/requirements.txt
django==4.2.15
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
profi_log==0.2.1